test(components-js): move e2e default dom test | sas
